Attempting to refactor messy python code
