+++ python/scraper/scraper.py
@@ -1,7 +1,71 @@
 from bs4 import BeautifulSoup
-from colorama import *
 import requests
 import re
+
+from utils import fetch_url
+from utils import base_url
+
+
+def get_comic_data(ch_link: str, pages_amount: int = 1):
+    # grab the first comic of a chapter
+    if pages_amount != 1:
+        url_set = f"{ch_link}page/{pages_amount}/"
+    else:
+        url_set = ch_link
+    gc_url = fetch_url(url_set)
+    comic_item = BeautifulSoup(gc_url.text, "html.parser").find_all(
+        "article", id=re.compile("^post-"))
+
+    first_comic = comic_item[-1]
+    comic_link: str = first_comic.find("a")["href"]
+    comic_date: str = first_comic.find(
+        "span", class_=re.compile("^mh-meta-date")).get_text()
+
+    return {
+        'comic_link': comic_link,
+        'comic_date': comic_date,
+    }
+
+
+def grab_chapters_comic():
+    # grabs the first comic for each chapter
+
+    print("Grabbing the first comic for each chapter...")
+    first_chapter_comics = dict()
+
+    # grab the chapters
+    archive_url = fetch_url(f"{base_url}/archive/")
+    chapter_dropdown = BeautifulSoup(
+        archive_url.text, "html.parser").find_all("option", class_="level-0")
+
+    for chapters in chapter_dropdown:
+        name_parse = re.sub("^(-|\d)(\d\d).\s", "", chapters.get_text())
+        ch_link, ch_name = chapters["value"], name_parse
+
+        chapter_url = fetch_url(ch_link)
+        print(ch_link)
+        chapter_soup = BeautifulSoup(chapter_url.text, "html.parser")
+        # print(chapter_soup)
+
+        pagination = chapter_soup.find(
+            "div", class_=re.compile("^mh-loop-pagination"))
+        pag_total = 1
+        # if there is more than 1 page, get the amount of pages
+        if pagination != None:
+            pag_total = int(pagination.find_all(
+                "a", class_="page-numbers")[-2].get_text())
+
+        # grab the first comic from the chapter
+        first_comic = get_comic_data(ch_link, pag_total)
+        first_comic.update({"ch_name": ch_name})
+
+        print({first_comic['comic_link'].split("/")[-2]: first_comic})
+        first_chapter_comics.update({
+            # uses the link to grab the comic tittle
+            first_comic['comic_link'].split("/")[-2]: first_comic
+        })
+
+    return first_chapter_comics
 
 
 def scrape_comic(url, year, index, characters_db):
@@ -43,8 +107,7 @@ def scrape_comic(url, year, index, characters_db):
         character_str = ",".join(characters)
 
     else:
-        print(
-            f"{Fore.BLACK}{Back.LIGHTWHITE_EX}{Style.BRIGHT}{url} is a guest comics{Style.RESET_ALL}")
+        print(f"{url} is a guest comics")
 
     return {
         "key_name": f"{year}:{link_title}",
